feat: Add Form component and configure routing
- Moved form code to a separate Form component
- Created FormPage with a title and the Form component
- Set up routing using react-router-dom to handle navigation
- Updated App component to include routes for the home page and form page
