+++ package.json
@@ -12,7 +12,9 @@
   "dependencies": {
     "lucide-react": "^0.427.0",
     "react": "^18.3.1",
-    "react-dom": "^18.3.1"
+    "react-dom": "^18.3.1",
+    "react-router-dom": "^6.26.0",
+    "sweetalert2": "^11.12.4"
   },
   "devDependencies": {
     "@eslint/js": "^9.8.0",
